requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
sqlalchemy==2.0.35
aiosqlite==0.20.0
httpx==0.27.2
python-telegram-bot==21.6
neonize
openai==1.51.0
anthropic==0.34.2
google-generativeai==0.8.2
google-genai>=1.0.0
boto3==1.35.30
google-cloud-aiplatform==1.69.0
pydantic==2.9.2
rich>=13.0.0
questionary>=2.0.0
